Bug fixes. PublishVideo, UpdateVideo, DeleteVideo controllers added.

+++ src/controllers/video.controller.js
@@ -16,8 +16,48 @@ const getAllVideos = asyncHandler(async (req, res) => {
 });
 
 const publishAVideo = asyncHandler(async (req, res) => {
+    //* Sanity check of title and description
     const { title, description } = req.body;
-    // TODO: get video, upload to cloudinary, create video
+    if (!title || !description)
+        throw new ApiError(400, "All fields are required.");
+
+    //* Get the uploaded file local path and perform sanity check
+    const videoLocalFilePath = req.files?.video;
+    const thumbnailLocalPath = req.files?.thumbnail;
+    if (!videoLocalFilePath || !thumbnailLocalPath)
+        throw new ApiError(400, "All files are required.");
+
+    //* Upload the files into cloudinary and check if upload was successfull
+    const video = await uploadOnCloudinary(videoLocalFilePath);
+    const thumbnail = await uploadOnCloudinary(thumbnailLocalPath);
+    if (!video || !thumbnail)
+        throw new ApiError(
+            400,
+            "Something went wrong while uploading the file."
+        );
+
+    //* Create new video object and upload into Database
+    const newVideo = {
+        videoFile: video.url,
+        thumbnail: thumbnail.url,
+        title,
+        description,
+        duration: 1,
+        isPublished: true,
+        owner: req.user?._id,
+    };
+    const result = await Video.create(newVideo);
+    if (!result)
+        throw new ApiError(
+            400,
+            "Something went wrong while posting the video."
+        );
+
+    return res
+        .status(201)
+        .json(
+            new ApiResponse(201, result, "Video has been posted successfully.")
+        );
 });
 
 const getVideoById = asyncHandler(async (req, res) => {
@@ -75,6 +115,46 @@ const getVideoById = asyncHandler(async (req, res) => {
 const updateVideo = asyncHandler(async (req, res) => {
     const { videoId } = req.params;
     //TODO: update video details like title, description, thumbnail
+
+    //* Sanity check
+    const { title, description } = req.body;
+    if (!title || !description)
+        throw new ApiError(400, "All fields are required.");
+    const isVideoIdValid = sanityCheck(videoId);
+    if (!isVideoIdValid) throw new ApiError(400, "Video not found.");
+
+    const thumbnailLocalPath = req.file?.thumbnail;
+    let thumbnailUpload = null;
+    if (thumbnailLocalPath) {
+        thumbnailUpload = await uploadOnCloudinary(thumbnailLocalPath);
+    }
+
+    //* Check if video exists
+    const videoDetails = await Video.findById(videoId);
+    if (!videoDetails) throw new ApiError(400, "Video not found.");
+
+    const newVideoDetails = await Video.findByIdAndUpdate(
+        videoId,
+        {
+            $set: {
+                title,
+                description,
+                thumbnail: thumbnailUpload
+                    ? thumbnailUpload.url
+                    : videoDetails.thumbnail,
+            },
+        },
+        { new: true }
+    );
+    return res
+        .status(200)
+        .json(
+            new ApiResponse(
+                200,
+                newVideoDetails,
+                "Video details updated successfully."
+            )
+        );
 });
 
 const deleteVideo = asyncHandler(async (req, res) => {
@@ -82,9 +162,18 @@ const deleteVideo = asyncHandler(async (req, res) => {
     const isVideoIdValid = sanityCheck(videoId);
     if (!isVideoIdValid) throw new ApiError(400, "Video not found.");
 
+    //* Check if video exists or not
+    const video = await Video.findById(videoId);
+    if (!video) throw new ApiError(400, "Video not found.");
+
+    //* Check if current user is the owner of the video
+    if (video.owner !== req.user?._id)
+        throw new ApiError(401, "User unuthorized.");
+
     const videoDetails = await Video.findByIdAndDelete(videoId);
     if (!videoDetails) throw new ApiError(400, "Video not found.");
 
+    //* Delete existing thumbnail and video from cloudinary
     const currentVideoFile = getPublicId(videoDetails.videoFile);
     const currentThumbnail = getPublicId(videoDetails.thumbnail);
     await deleteFromCloudinary(currentVideoFile);
@@ -105,6 +194,10 @@ const togglePublishStatus = asyncHandler(async (req, res) => {
     //* Check if video exists or not
     const video = await Video.findById(videoId);
     if (!video) throw new ApiError(400, "Video not found.");
+
+    //* Check if current user is the owner of the video
+    if (video.owner !== req.user?._id)
+        throw new ApiError(401, "User unuthorized.");
 
     //* Toggle the isPublished value and save the info in database
     video.isPublished = !video.isPublished;
